+++ sync-for-payables/src/sdk/models/webhooks/index.ts
@@ -4,3 +4,6 @@
 
 export * from "./clientratelimitreached";
 export * from "./clientratelimitreset";
+export * from "./payablesbillcreated";
+export * from "./payablesbilldeleted";
+export * from "./payablesbillupdated";